requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
command-not-found==0.3
community==1.0.0b1
confection==0.1.5
cryptography==3.4.8
cymem==2.0.11
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.20.0
Flask==3.1.2
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
gyp==0.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
IMAPClient==3.1.0
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
langchain==1.1.3
langchain-core==1.1.2
langchain-openai==1.1.1
langcodes==3.5.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
log_color==2.1.0
lxml==6.0.2
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
meson==1.10.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.13
narwhals==2.13.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
openai==2.9.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
preshed==3.0.10
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==5.4.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.16.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.52.1
sympy==1.14.0
systemd-python==234
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
uuid_utils==0.12.0
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==1.0.0
zstandard==0.25.0
